Make throw throw when push is unhandled
This is the first time I got Repeater.prototype.throw working with
current tests. The strategy I’m using is that we chain this.pending in
the push method rather than in the unwrap method. This refactoring of
when we chain promises broke a lot of code initially, but I managed to
get it working with the current test suite. I’m not sure about the
changes to types, particularly passing promises to the buffer instead of
naked values when the pushed values are not promises. I wonder if there
is a hybrid approach where we chain promises in push but still pass
unwrapped values into the pushQueue/buffer.